test_simple: cast to avoid `-Wsign-conversion`
```
tests/test_simple.c:84:56: error: operand of ? changes signedness: 'int' to 'unsigned long' [clang-diagnostic-sign-conversion]
   84 |         gcry_mpi_t f = gcry_mpi_set_ui(NULL, fi >= 0 ? fi : -fi);
      |                        ~~~~~~~~~~~~~~~                 ^~
tests/test_simple.c:84:61: error: operand of ? changes signedness: 'int' to 'unsigned long' [clang-diagnostic-sign-conversion]
   84 |         gcry_mpi_t f = gcry_mpi_set_ui(NULL, fi >= 0 ? fi : -fi);
      |                        ~~~~~~~~~~~~~~~                      ^~~
tests/test_simple.c:85:46: error: implicit conversion changes signedness: 'int' to 'unsigned long' [clang-diagnostic-sign-conversion]
   85 |         gcry_mpi_t p = gcry_mpi_set_ui(NULL, atoi(tests[i].p));
      |                        ~~~~~~~~~~~~~~~       ^~~~~~~~~~~~~~~~
```
Ref: https://github.com/libssh2/libssh2/actions/runs/30043871895/job/89330114459#step:24:124

Follow-up to 76d30e5d984cdc88ed802aafc12a34132f370ed8 #2417
Follow-up to fcf830788d0c8dcd6bed7dcd532506aabf196523 #2239

+++ tests/test_simple.c
@@ -81,8 +81,8 @@ static int test_ssh2_dh_validate(void)
         int got;
 #ifdef LIBSSH2_LIBGCRYPT
         int fi = atoi(tests[i].f);
-        gcry_mpi_t f = gcry_mpi_set_ui(NULL, fi >= 0 ? fi : -fi);
-        gcry_mpi_t p = gcry_mpi_set_ui(NULL, atoi(tests[i].p));
+        gcry_mpi_t f = gcry_mpi_set_ui(NULL, (unsigned long)(fi >= 0 ? fi : -fi));
+        gcry_mpi_t p = gcry_mpi_set_ui(NULL, (unsigned long)atoi(tests[i].p));
         if(tests[i].f[0] == '-')
             gcry_mpi_neg(f, f);
         got = ssh2_dh_validate(f, p);
